refactor(ui): consolidate shared CSS, extract utilities, and split Templates page
- Create shared.module.css with reusable form, table, and layout classes
- Extract format utilities (tryParseJson, parseLabels, formatBytes, etc.) into lib/format.ts
- Split Templates.tsx (1200+ lines) into TemplatesTab, ProfilesTab, PlaygroundTab, and profileUtils
- Create shared DeleteConfirmModal, DetailGrid, and LabelsInput components
- Replace duplicated per-page CSS with shared module classes across 12 page components
- Reduce total CSS by ~250 lines and page component code by ~1600 lines
