+++ ui/src/pages/RetentionPolicies.module.css
@@ -1,29 +1,7 @@
 @reference "../index.css";
 
-.filterBar {
-  @apply flex flex-wrap gap-3 mb-4;
-}
-
-.actionsCell {
-  @apply flex items-center gap-1;
-}
-
 .ttlValue {
   @apply font-mono text-xs;
-  color: var(--text-secondary);
-}
-
-/* Create / Edit modal */
-.formGrid {
-  @apply grid grid-cols-2 gap-4;
-}
-
-.formSection {
-  @apply space-y-4;
-}
-
-.textareaLabel {
-  @apply block text-sm font-medium mb-1;
   color: var(--text-secondary);
 }
 
@@ -43,38 +21,10 @@
   @apply space-y-3 text-sm;
 }
 
-.detailRow {
-  @apply flex justify-between py-1 border-b;
-  border-color: var(--border-default);
-}
-
-.detailLabel {
-  @apply text-gray-500;
-}
-
-.detailValue {
-  @apply font-mono text-xs;
-}
-
 .detailValueWrap {
   @apply font-mono text-xs text-right max-w-xs break-all;
 }
 
 .sectionTitle {
   @apply text-sm font-medium mb-2;
-}
-
-.actionButtons {
-  @apply mt-6 flex gap-2;
-}
-
-/* Delete confirmation */
-.deleteWarning {
-  @apply text-sm mb-4;
-  color: var(--text-secondary);
-}
-
-.deleteName {
-  @apply font-semibold;
-  color: var(--text-primary);
 }
