+++ ui/src/pages/Templates.module.css
@@ -10,16 +10,6 @@
   @apply mb-6;
 }
 
-/* Filter bar */
-.filterBar {
-  @apply flex flex-wrap gap-3 mb-4;
-}
-
-/* Table row actions */
-.actionsCell {
-  @apply flex items-center gap-1;
-}
-
 /* Cells */
 .nameCell {
   @apply font-mono text-sm;
@@ -37,19 +27,6 @@
 }
 
 /* Create / Edit modal form */
-.formSection {
-  @apply space-y-4;
-}
-
-.formGrid {
-  @apply grid grid-cols-2 gap-4;
-}
-
-.textareaLabel {
-  @apply block text-sm font-medium mb-1;
-  color: var(--text-secondary);
-}
-
 .textarea {
   @apply w-full rounded-md border px-3 py-2 font-mono text-sm focus:outline-none focus:ring-2 focus:ring-primary-400/20 focus:border-primary-400;
   background: var(--bg-surface);
@@ -113,19 +90,6 @@
   @apply space-y-3 text-sm;
 }
 
-.detailRow {
-  @apply flex justify-between py-1 border-b;
-  border-color: var(--border-default);
-}
-
-.detailLabel {
-  color: var(--text-muted);
-}
-
-.detailValue {
-  @apply font-mono text-xs;
-}
-
 .detailValueWrap {
   @apply font-mono text-xs text-right max-w-xs break-all;
 }
@@ -137,15 +101,6 @@
   color: var(--text-default);
   white-space: pre-wrap;
   word-break: break-word;
-}
-
-.sectionTitle {
-  @apply text-sm font-medium mb-2 mt-4;
-  color: var(--text-secondary);
-}
-
-.actionButtons {
-  @apply mt-6 flex gap-2;
 }
 
 /* Fields display in drawer */
@@ -173,17 +128,6 @@
   background: var(--bg-surface-secondary);
   color: var(--text-muted);
   border: 1px solid var(--border-default);
-}
-
-/* Delete confirmation */
-.deleteWarning {
-  @apply text-sm mb-4;
-  color: var(--text-secondary);
-}
-
-.deleteName {
-  @apply font-semibold;
-  color: var(--text-primary);
 }
 
 /* ---- Playground Tab ---- */
