+++ ui/src/pages/WasmPlugins.module.css
@@ -1,10 +1,6 @@
 @reference "../index.css";
 
 /* Plugin list table */
-.actionsCell {
-  @apply flex items-center gap-1;
-}
-
 .nameCell {
   @apply font-mono text-sm;
   color: var(--text-default);
@@ -26,14 +22,6 @@
 }
 
 /* Register modal */
-.formSection {
-  @apply space-y-4;
-}
-
-.formGrid {
-  @apply grid grid-cols-2 gap-4;
-}
-
 .fileInput {
   @apply w-full text-sm border rounded-md px-3 py-2 cursor-pointer;
   background: var(--bg-surface);
@@ -42,11 +30,6 @@
 }
 
 .fileLabel {
-  @apply block text-sm font-medium mb-1;
-  color: var(--text-secondary);
-}
-
-.textareaLabel {
   @apply block text-sm font-medium mb-1;
   color: var(--text-secondary);
 }
@@ -63,40 +46,13 @@
   @apply space-y-3 text-sm;
 }
 
-.detailRow {
-  @apply flex justify-between py-1 border-b;
-  border-color: var(--border-default);
-}
-
-.detailLabel {
-  color: var(--text-muted);
-}
-
-.detailValue {
-  @apply font-mono text-xs;
-}
-
 .detailValueWrap {
   @apply font-mono text-xs text-right max-w-xs break-all;
 }
 
-.sectionTitle {
-  @apply text-sm font-medium mb-2 mt-4;
-}
-
-.actionButtons {
-  @apply mt-6 flex gap-2;
-}
-
-/* Delete confirmation */
-.deleteWarning {
-  @apply text-sm mb-4;
-  color: var(--text-secondary);
-}
-
-.deleteName {
-  @apply font-semibold;
-  color: var(--text-primary);
+/* Test field spacing */
+.testFieldSpacing {
+  @apply mt-3;
 }
 
 /* Test panel */
